+++ pennylane/ops/op_math/decompositions/__init__.py
@@ -15,6 +15,5 @@ r"""This module contains decompositions for (numerically-specified) arbitrary
 unitary operations into sequences of elementary operations.
 """
 
-from .unitary_decompositions import one_qubit_decomposition
+from .unitary_decompositions import one_qubit_decomposition, two_qubit_decomposition
 from .solovay_kitaev import sk_decomposition
-from .two_qubit_unitary import two_qubit_decomposition
